Odyssey stage 1
All events in there, not all working, to V30 spec
MCMRList changed to generational dictionary, seems to work with comm/mats
Its not ItemDate for ModuleData
Some basic suit system in there but embryonic
ulongs for fd ids for ships

+++ EliteDangerous/JournalEvents/JournalSetUserShipName.cs
@@ -25,14 +25,14 @@ namespace EliteDangerousCore.JournalEvents
         {
             ShipFD = JournalFieldNaming.NormaliseFDShipName(evt["Ship"].Str());
             Ship = JournalFieldNaming.GetBetterShipName(ShipFD);
-            ShipID = evt["ShipID"].Int();
+            ShipID = evt["ShipID"].ULong();
             ShipName = evt["UserShipName"].Str();// name to match LoadGame
             ShipIdent = evt["UserShipId"].Str();     // name to match LoadGame
         }
 
         public string Ship { get; set; }
         public string ShipFD { get; set; }
-        public int ShipID { get; set; }
+        public ulong ShipID { get; set; }
         public string ShipName { get; set; }
         public string ShipIdent { get; set; }
